requirements.txt
pydantic==1.10.12
httpx
pyyaml
